requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
PyMySQL==1.1.0
python-dotenv==1.0.0
openpyxl==3.1.2
reportlab==4.0.8
